fix(1.1.0): поле ввода, сессии (float тип), slash-команды, CRUD ключей, связи сущностей
ФИКСЫ:
- Поле ввода не очищалось: очищаем ДО invoke; send_message_cmd fire-and-
  forget (стриминг через события, не блокирует UI)
- Сессии пусты: Hermes state.db хранит started_at как REAL (float), Rust
  ожидал i64 → query_map падал. Все time-поля SessionSummary/SessionMessage/
  SearchResult → f64.
- Голос: MediaRecorder primary для live+note; window.focus() перед getUserMedia
  для видимости permission-диалога.

ФИЧИ:
- Slash-команды (/model /clear /compact /profile /help /tasks /skills /cost):
  popup-меню при вводе /, навигация стрелками, выбор Enter/Tab.
- API ключи: delete (remove_credential_pool_entry_cmd), provider select из всех
  ~30 провайдеров Hermes (PROVIDERS.options).
- Выбор модели: provider select расширен до полного списка провайдеров.
- Связи Задача→Проект→Цель: projects.goal_id (миграция ALTER TABLE), выбор цели
  при создании проекта, отображение связи в карточке.

Версия 1.1.0. tsc+cargo+build чисто.

+++ src-tauri/src/sessions.rs
@@ -15,8 +15,8 @@ use crate::config::profile_home;
 pub struct SessionSummary {
     pub id: String,
     pub source: String,
-    pub started_at: i64,
-    pub ended_at: Option<i64>,
+    pub started_at: f64,
+    pub ended_at: Option<f64>,
     pub message_count: i64,
     pub model: String,
     pub title: Option<String>,
@@ -28,14 +28,14 @@ pub struct SessionMessage {
     pub id: i64,
     pub role: String,
     pub content: String,
-    pub timestamp: i64,
+    pub timestamp: f64,
 }
 
 #[derive(Debug, Clone, Serialize)]
 pub struct SearchResult {
     pub session_id: String,
     pub title: Option<String>,
-    pub started_at: i64,
+    pub started_at: f64,
     pub source: String,
     pub message_count: i64,
     pub model: String,
